Update IP detection routine
IP detection was failing for libvirt when more than 1 nic was
configured.  This patch makes detection the same for all hosts.

The detection selects the last nic's IPv4 Address from all nics
that are up.

The custom detection for VirtualBox is no longer needed as we now
force add a second nic.